Change type from int to float64 for math funcs

+++ add.go
@@ -1,6 +1,6 @@
 package goblin
 
 // Adds two numbers.
-func Add(augend, addend int) int {
+func Add(augend, addend float64) float64 {
 	return augend + addend
 }
