+++ multiply.go
@@ -1,6 +1,6 @@
 package goblin
 
 // Multiply two numbers.
-func Multiply(multiplier, multiplicand int) int {
+func Multiply(multiplier, multiplicand float64) float64 {
 	return multiplier * multiplicand
 }
