+++ subtract.go
@@ -1,6 +1,6 @@
 package goblin
 
 // Subtract two numbers.
-func Subtract(minuend, subtrahend int) int {
+func Subtract(minuend, subtrahend float64) float64 {
 	return minuend - subtrahend
 }
