+++ subtract_test.go
@@ -7,9 +7,9 @@ import (
 )
 
 func TestSubtract(t *testing.T) {
-	difference := Subtract(6, 4)
-	assert.Equal(t, difference, 2)
+	difference := Subtract(float64(6), float64(4))
+	assert.Equal(t, difference, float64(2))
 
-	difference = Subtract(4, 6)
-	assert.Equal(t, difference, -2)
+	difference = Subtract(float64(4), float64(6))
+	assert.Equal(t, difference, float64(-2))
 }
